resume-match-agent/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
python-multipart==0.0.9
PyMuPDF==1.23.26
openai==1.30.1
python-dotenv==1.0.1
httpx==0.27.0 